ml-backend/requirements.txt
label-studio-ml
ultralytics
pillow
requests
redis
rq
gunicorn
pyyaml